fix build break
SVN r8529 |2008-04-11 21:41:10 +0000

+++ src/org/labkey/test/Runner.java
@@ -369,7 +369,7 @@ public class Runner extends TestSuite
 
     private static void writeTimeReport()
     {
-        long width = 60;
+        int width = 60;
         long total = 0;
         System.out.println("======================= Time Report ========================");
